% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_download_images.R
\name{img_download_images}
\alias{img_download_images}
\title{Download images from urls.}
\usage{
img_download_images(link, dir, prefix, scale_border = T)
}
\arguments{
\item{link}{URL of image}

\item{dir}{Directory to download images}

\item{prefix}{File name prefix for downloaded images}

\item{scale_border}{Boolean, scale images and add border}
}
\value{
Image files
}
\description{
Download images from urls.
}
