% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_get_gurls.R
\name{img_get_gurls}
\alias{img_get_gurls}
\title{Get links to Google images per some Google search}
\usage{
img_get_gurls(x)
}
\arguments{
\item{x}{Search term}
}
\value{
A data frame of URLs
}
\description{
Get links to Google images per some Google search
}
