% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_build_collage.R
\name{img_build_collage}
\alias{img_build_collage}
\title{Aggregate images as collage.}
\usage{
img_build_collage(dir, dimx, dimy, prefix = NULL)
}
\arguments{
\item{dir}{Directory of images}

\item{dimx}{Collage dimension on x-axis}

\item{dimy}{Collage dimension on y-axis}

\item{prefix}{File name prefix of images to include in collage}
}
\value{
An image
}
\description{
Aggregate images as collage.
}
